services/backend/requirements.txt
fastapi[all]==0.103.0
uvicorn[standard]>=0.23.0,<0.24.0
gunicorn>=21.0.0,<22.0.0
pydantic>=2.3.0,<2.4.0
sqlalchemy>=2.0.21,<2.1.0
psycopg2-binary>=2.9.5,<2.10.0
alembic>=1.12.0,<1.13.0
python-jose[cryptography]
passlib[bcrypt]